Refactor expiring auth module structure
